- 重构脚本 - 修改SnowflakeComponent为kotlin

+++ build.sh
@@ -1,6 +1,14 @@
-rm -rf ./bin
+version=1.0.0-SNAPSHOT
+
+if [ -d './bin' ]; then
+  rm -rf ./bin
+fi
+gradle clean
 gradle build
-sleep 1
+sleep 3
+mkdir ./bin
+cp ./server/build/libs/tinkertian-snowflake-server-${version}.jar ./bin/.
 cp ./scripts/sf.sh ./bin/.
 chmod u+x ./bin/sf.sh
 rm -rf ./gradle
+rm -rf ./build
